requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.8
dj-database-url==0.5.0
Django==3.2.6
django-extensions==3.1.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
docker==5.0.2
ecdsa==0.17.0
gunicorn==20.1.0
hvac==0.11.0
idna==3.2
kombu==5.1.0
libvirt-python==7.7.0
prompt-toolkit==3.0.20
psycopg2==2.9.1
py-cpuinfo==8.0.0
pycparser==2.20
PyJWT==2.1.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
sshpubkeys==3.3.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
websocket-client==1.2.1
xmltodict==0.12.0
